ENH: move nrdtrisd and nrdtrimn to stats.h
